table to string util test
